+++ lib/modules/storage/web/settings.ex
@@ -33,7 +33,6 @@ defmodule PhoenixKitWeb.Live.Modules.Storage.Settings do
     # Load storage settings from database (using basic function to avoid cache issues)
     redundancy_copies = Settings.get_setting("storage_redundancy_copies", "1")
     auto_generate_variants = Settings.get_setting("storage_auto_generate_variants", "true")
-    default_bucket_uuid = Settings.get_setting("storage_default_bucket_uuid", nil)
     max_upload_size_mb = Settings.get_setting("storage_max_upload_size_mb", "500")
     tile_generation_enabled = Settings.get_setting("storage_tile_generation_enabled", "false")
 
@@ -71,7 +70,6 @@ defmodule PhoenixKitWeb.Live.Modules.Storage.Settings do
       |> assign(:auto_generate_variants, auto_generate_variants == "true")
       |> assign(:tile_generation_enabled, tile_generation_enabled == "true")
       |> assign(:annotated_thumbnails_enabled, annotated_thumbnails_enabled == "true")
-      |> assign(:default_bucket_uuid, default_bucket_uuid)
       |> assign(:active_buckets_count, active_buckets)
       |> assign(:max_redundancy, max_redundancy)
       |> assign(:form_redundancy, form_redundancy)
@@ -340,24 +338,6 @@ defmodule PhoenixKitWeb.Live.Modules.Storage.Settings do
     end
   end
 
-  def handle_event("update_default_bucket", %{"bucket_uuid" => bucket_uuid}, socket) do
-    new_value = if bucket_uuid == "", do: nil, else: bucket_uuid
-
-    case Settings.update_setting("storage_default_bucket_uuid", new_value) do
-      {:ok, _setting} ->
-        socket =
-          socket
-          |> assign(:default_bucket_uuid, new_value)
-          |> put_flash(:info, gettext("Default bucket updated"))
-
-        {:noreply, socket}
-
-      {:error, _changeset} ->
-        socket = put_flash(socket, :error, gettext("Failed to update default bucket"))
-        {:noreply, socket}
-    end
-  end
-
   def handle_event("toggle_bucket", %{"id" => bucket_uuid}, socket) do
     case Storage.get_bucket(bucket_uuid) do
       nil ->
@@ -515,7 +495,6 @@ defmodule PhoenixKitWeb.Live.Modules.Storage.Settings do
     # Reload storage settings
     redundancy_copies = Settings.get_setting("storage_redundancy_copies", "1")
     auto_generate_variants = Settings.get_setting("storage_auto_generate_variants", "true")
-    default_bucket_uuid = Settings.get_setting("storage_default_bucket_uuid", nil)
     max_upload_size_mb = Settings.get_setting("storage_max_upload_size_mb", "500")
 
     # Recalculate max redundancy
@@ -529,7 +508,6 @@ defmodule PhoenixKitWeb.Live.Modules.Storage.Settings do
     |> assign(:bucket_file_counts, bucket_file_counts)
     |> assign(:redundancy_copies, current_redundancy)
     |> assign(:auto_generate_variants, auto_generate_variants == "true")
-    |> assign(:default_bucket_uuid, default_bucket_uuid)
     |> assign(:active_buckets_count, active_buckets_count)
     |> assign(:max_redundancy, max_redundancy)
     |> assign(:form_redundancy, current_redundancy)